+++ src/main/java/pl/plajer/murdermystery/arena/ArenaUtils.java
@@ -71,7 +71,7 @@ public class ArenaUtils {
         ChatManager.colorMessage("In-Game.Messages.Game-End-Messages.Subtitles.Murderer-Stopped", murderer), 5, 40, 5);
     }
     //we must call it ticks later due to instant respawn bug
-    Bukkit.getScheduler().runTaskLater(plugin, () -> ArenaManager.stopGame(false, arena), 5);
+    Bukkit.getScheduler().runTaskLater(plugin, () -> ArenaManager.stopGame(false, arena), 10);
   }
 
   public static void addScore(User user, ScoreAction action, int amount) {