+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+### 1.6.0 Release
+* Fixed spectators preventing arrows
+* Fixed editor value "spawn gold time"
+* Added bungeemode Arena shuffler
+* Added option to disable separate chat
+* Added 1.16.x compatibility 
+* Added possibility to change table name of mysql database
+* Removed stats save on quitevent as it wasn't necessary
+* Fixed different sword cooldown time than 5 seconds
+* Fixed bungeemanager enable null text error
+* Optimized murderer leave before they got the sword
+* Optimized the death respawn location 
+* Fixed prayer multi gold
+
+### 1.5.5 Hotfix
+* Fixed language list getting
+
 ### 1.5.4 Release
 * Added new config values to change arrow amount of fake detective and bow on gold
 * Fixed properties getting on language list